Toward respecting DiffStrategy
