+++ pkg/tfshim/sdk-v2/provider.go
@@ -88,34 +88,6 @@ func (p v2Provider) Configure(c shim.ResourceConfig) error {
 	return errors(p.tf.Configure(context.TODO(), configFromShim(c)))
 }
 
-func (p v2Provider) Diff(t string, s shim.InstanceState, c shim.ResourceConfig) (shim.InstanceDiff, error) {
-	if c == nil {
-		return diffToShim(&terraform.InstanceDiff{Destroy: true}), nil
-	}
-	r, ok := p.tf.ResourcesMap[t]
-	if !ok {
-		return nil, fmt.Errorf("unknown resource %v", t)
-	}
-
-	config, state := configFromShim(c), stateFromShim(s)
-	rawConfig := makeResourceRawConfig(config, r)
-	if state != nil {
-		state.RawConfig = rawConfig
-	}
-
-	state, err := upgradeResourceState(p.tf, r, state)
-	if err != nil {
-		return nil, fmt.Errorf("failed to upgrade resource state: %w", err)
-	}
-
-	diff, err := simpleDiff(context.TODO(), r, state, config, rawConfig, p.tf.Meta())
-	if diff != nil {
-		diff.RawConfig = rawConfig
-	}
-
-	return diffToShim(diff), err
-}
-
 func (p v2Provider) Apply(t string, s shim.InstanceState, d shim.InstanceDiff) (shim.InstanceState, error) {
 	r, ok := p.tf.ResourcesMap[t]
 	if !ok {
